requirements.txt
asgiref==3.2.7
betterpath==0.2.2
click==7.1.2
Django==3.0.6
django-crispy-forms==1.9.1
django-model-utils==4.0.0
l==0.11.0
Pillow==7.1.2
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
vcversioner==2.16.0.0
zope.interface==5.1.0
